+++ scripts/module.js
@@ -35,6 +35,7 @@ import {vaeEffectDescription, vaeTempItemButton} from './integrations/vae.js';
 import {checkUpdate} from './update.js';
 import {firearm} from './macros/mechanics/firearm.js';
 import {templateMacroTitleBarButton} from './integrations/templateMacro.js';
+import {addActions} from './macros/actions/token.js';
 export let socket;
 Hooks.once('init', async function() {
     registerSettings();
@@ -191,6 +192,7 @@ Hooks.once('ready', async function() {
     if (game.settings.get('chris-premades', 'Critical Role Firearm Support')) firearm.setup(true);
     if (game.settings.get('chris-premades', 'Booming Blade')) Hooks.on('updateToken', macros.boomingBlade.moved);
     if (game.settings.get('chris-premades', 'Build A Bonus Overlapping Effects')) Hooks.on('babonus.filterBonuses', buildABonus.overlappingEffects);
+    Hooks.on('createToken', addActions);
 });
 let dev = {
     'setCompendiumItemInfo': setCompendiumItemInfo,